Split out data reading
For now, data reading simply removes punctuation and splits the file
into its words list, but in the future it will get more complex. We
will want to preserve punctuation and paragraph structure. As such,
split the action into its own method where implementation iteration
can be done in the future without affecting the rest of the code.