Extend to iterate over all .txt files in a directory
I gave the OHHLA scraper some TLC and in doing so, split the results
to one text file per artist containing all of their lyrics. This
gave a lot of benefits on the scraper side, but means that this
project no longer works over the same inputs/outputs.

As such, extend this project to consume lyrics spread out over a
number of text files all contained in a single directory. Further, due
to the potential volume of data, stream the results via a generator.